Feat/board api

+++ src/main/java/com/b3/ddarelro/domain/board/service/BoardService.java
@@ -1,13 +1,18 @@
 package com.b3.ddarelro.domain.board.service;
 
 import com.b3.ddarelro.domain.board.dto.request.BoardCreateReq;
+import com.b3.ddarelro.domain.board.dto.request.BoardDropReq;
+import com.b3.ddarelro.domain.board.dto.request.BoardInviteReq;
+import com.b3.ddarelro.domain.board.dto.request.BoardLeaveReq;
 import com.b3.ddarelro.domain.board.dto.request.BoardUpdateReq;
 import com.b3.ddarelro.domain.board.dto.response.BoardCreateRes;
 import com.b3.ddarelro.domain.board.dto.response.BoardDeleteRes;
 import com.b3.ddarelro.domain.board.dto.response.BoardDetailRes;
 import com.b3.ddarelro.domain.board.dto.response.BoardDropRes;
 import com.b3.ddarelro.domain.board.dto.response.BoardInviteRes;
+import com.b3.ddarelro.domain.board.dto.response.BoardLeaveRes;
 import com.b3.ddarelro.domain.board.dto.response.BoardPriviewRes;
+import com.b3.ddarelro.domain.board.dto.response.BoardRestoreRes;
 import com.b3.ddarelro.domain.board.dto.response.BoardUpdateRes;
 
 import com.b3.ddarelro.domain.board.entity.Board;
@@ -15,12 +20,14 @@ import com.b3.ddarelro.domain.board.exception.BoardErrorCode;
 import com.b3.ddarelro.domain.board.repository.BoardRepository;
 import com.b3.ddarelro.domain.column.service.ColumnService;
 import com.b3.ddarelro.domain.user.entity.User;
+import com.b3.ddarelro.domain.user.service.UserService;
 import com.b3.ddarelro.domain.userboard.entity.BoardAuthority;
 import com.b3.ddarelro.domain.userboard.entity.UserBoard;
 import com.b3.ddarelro.domain.userboard.repository.UserBoardRepository;
 import com.b3.ddarelro.global.exception.GlobalException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -85,18 +92,32 @@ public class BoardService {
         //board 작성자가 맞는지확인
         validteUserAuthority(founddUser,foundBoard);
 
-        foundBoard.deleteBoardState(true);
+        foundBoard.updateBoardState(true);
         //columnService.deleteAllColumn(boardId);
         return new BoardDeleteRes("삭제가 완료되었습니다.");
     }
 
-    @Transactional(readOnly = true)
-    public List<BoardPriviewRes> getBoardList() {
+    public BoardRestoreRes restoreBoard(Long userId, Long boardId){
 
-        List<Board> boardList =  boardRepository.findAll();
+        User founddUser = userService.findUser(userId);
+        Board foundBoard = findBoard(boardId);
+
+        foundBoard.updateBoardState(false);
+
+        return new BoardRestoreRes(foundBoard);
+
+
+    }
+
+    @Transactional(readOnly = true)
+    public List<BoardPriviewRes> getBoardList(boolean isAsc,String sortBy) {
+
+        Sort.Direction direction = isAsc ? Sort.Direction.ASC : Sort.Direction.DESC;
+        Sort sort = Sort.by(direction, sortBy);
+
+        List<Board> boardList =  boardRepository.findAllByDeletedFalse(sort);
 
         return boardList.stream()
-            .filter(board -> board.getDeleted().equals(false))
             .map(BoardPriviewRes::new)
             .toList();
 
@@ -115,9 +136,9 @@ public class BoardService {
         return new BoardDetailRes(foundBoard);
     }
 
-    public BoardInviteRes inviteMember(Long userId, Long boardId, Long invitedUserId){
+    public BoardInviteRes inviteMember(Long userId, Long boardId, BoardInviteReq req){
 
-        validateInviteOwn(userId,invitedUserId); // 자기자신을 초대하는지 체크
+        Long invitedUserId = req.getUserId();
 
         User founddUser = userService.findUser(userId);
 
@@ -146,23 +167,41 @@ public class BoardService {
         return new BoardInviteRes("초대가 완료되었습니다.");
     }
 
-    public BoardDropRes dropMember(Long userId, Long boardId, Long dropUserId){
+    public BoardDropRes dropMember(Long userId, Long boardId, BoardDropReq req){
 
-        validateInviteOwn(userId,dropUserId); // 자기자신을 초대하는지 체크
+        Long dropUserId = req.getUserId();
+
+        if(dropUserId.equals(userId)){
+            throw new GlobalException(BoardErrorCode.FORBIDDEN_DROP_OWN);
+        }
+
+
         User founddUser = userService.findUser(userId);
         User dropUser = userService.findUser(dropUserId);
         Board foundBoard = findBoard(boardId);
         validteUserAuthority(founddUser,foundBoard);
 
-        validateMember(dropUser,foundBoard); //탈퇴할 사용자가 멤버에 있는지 확인
-
-        UserBoard userBoard = userBoardRepository.findByUserAndBoard(dropUser,foundBoard).get();
+        UserBoard userBoard = validateMember(dropUser,foundBoard); //탈퇴할 사용자가 멤버에 있는지 확인
 
         userBoardRepository.delete(userBoard);
 
 
         return new BoardDropRes("탈퇴처리되었습니다.");
     }
+
+    public BoardLeaveRes leaveBoard(Long userId, Long boardId, BoardLeaveReq req){
+        User founddUser = userService.findUser(userId);
+        Board foundBoard = findBoard(boardId);
+
+        UserBoard userBoard = validateMember(founddUser,foundBoard);
+
+        validedateLeaveMember(userBoard,req.getUserId());
+
+        userBoardRepository.delete(userBoard);
+
+        return new BoardLeaveRes("멤버에서 탈퇴하셨습니다.");
+    }
+
 
 
     public Board findBoard(Long id){
@@ -175,38 +214,42 @@ public class BoardService {
         return board;
     }
 
-    private void validteUserAuthority(User foundUser,Board foundBoard) {
+    public void validteUserAuthority(User foundUser,Board foundBoard) {
         // 해당 보드,유저를 가지고있는 userBoard확인 후 권한을 가지고있는 유저인지 체크
-        validateMember(foundUser,foundBoard);
-        UserBoard foundUserBoard = userBoardRepository.findByUserAndBoard(foundUser,foundBoard).get();
+        UserBoard foundUserBoard = validateMember(foundUser,foundBoard);
         if(foundUserBoard.getBoardAuthority() != BoardAuthority.ADMIN){
             throw new GlobalException(BoardErrorCode.UNAUTHORIZED_ACCESS_BOARD);
         }
     }
 
-    private void validateDeleted(Board board){ //삭제된 보드인지 확인
+    private void validateDeleted(Board board){ //이미 삭제된 보드인지 확인
         if(board.getDeleted().equals(true)){
             throw new GlobalException(BoardErrorCode.ALREADY_DELETED_BOARD);
         }
     }
 
-    private void validateMember(User user, Board board){ // 같은 user, board를 가지고있는 userboard인지 확인
-        if(!userBoardRepository.existsByUserAndBoard(user,board)){
-            throw new GlobalException(BoardErrorCode.NOT_BOARD_MEMBER);
-        }
+    public UserBoard validateMember(User user, Board board){ // 같은 user,board를 가지고있는 userboard가 있는지 확인
 
-    }
-
-
-    private void validateInviteOwn(Long userId, Long otherUserId){ //자기 자신을 초대 및 탈퇴 시키는지 확인
-        if(userId.equals(otherUserId)){
-            throw new GlobalException(BoardErrorCode.FORBIDDEN_INVITE_OWN);
-        }
+        return userBoardRepository.findByUserAndBoard(user,board)
+            .orElseThrow(()->new GlobalException(BoardErrorCode.NOT_BOARD_MEMBER));
     }
 
     private void validateExistedMember(User user, Board board){ // 이미 가입되어있느 멤버인지 확인
         if(userBoardRepository.existsByUserAndBoard(user,board)){
             throw new GlobalException(BoardErrorCode.ALREADY_BOARD_MEMBER);
+        }
+    }
+
+    private void validedateLeaveMember(UserBoard userBoard, Long userId) { //회원 자진 탈퇴시 검증메서드
+        if(userBoard.getBoardAuthority().equals(BoardAuthority.ADMIN)){
+            if(userId.equals(null)){
+                throw new GlobalException(BoardErrorCode.REQUIRED_NEW_BOARD_ADMIN); //팀장일경우 권한을 넘겨줘야합니다.
+            }
+
+            User DelegateeUser = userService.findUser(userId);
+            UserBoard updateUserBoard = validateMember(DelegateeUser,userBoard.getBoard()); //테스트 해줘야함
+            updateUserBoard.UpdateAuthority(BoardAuthority.ADMIN);
+
         }
     }
